logreader: When following a file, treat non-regular files specially
When following a file with follow-freq(), and it is not a regular
file, but something for which stat(2) always returns 0 size (such as
the /dev/klog character device on FreeBSD),
log_reader_io_follow_file() never went into the process input branch,
because the current position (0) was always at EOF (also 0).

To break out of this unfortunate situation, if the file to read from
is not a regular file, go into the process input loop anyway: at
worst, it will do an expensive noop.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>